Move file that's now only being used by the VS layer
